+++ plugins/glossary/src/Controller/DeleteController.php
@@ -5,7 +5,9 @@ namespace Plugin\Glossary\Controller;
 use App\Activity\ActivityService;
 use Plugin\Glossary\Activity\Messages\EntryDeleted;
 use Plugin\Glossary\Service\GlossaryService;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
 use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Component\Security\Http\Attribute\IsGranted;
 
@@ -28,9 +30,13 @@ final class DeleteController extends AbstractGlossaryController
         ]);
     }
 
-    #[Route('/{id}', name: 'app_plugin_glossary_delete', methods: ['GET'])]
-    public function delete(int $id): Response
+    #[Route('/{id}', name: 'app_plugin_glossary_delete', methods: ['POST'])]
+    public function delete(Request $request, int $id): Response
     {
+        if (!$this->isCsrfTokenValid('admin_glossary_delete' . $id, (string) $request->request->get('_token'))) {
+            throw new BadRequestHttpException('Invalid CSRF token.');
+        }
+
         $item = $this->service->get($id);
         $this->service->delete($id);
